requirements.txt
python-dotenv>=1.0.0
cryptography>=41.0.0
python-json-logger>=2.0.0